refactor: EditPostDialog 분리, features/dialog 추가
