+++ src/pages/PostsManagerPage.tsx
@@ -12,8 +12,11 @@ import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from ".
 import { useShowUserModal } from "../features/user/model/useShowUserModal"
 import UserProfileModal from "../features/user/ui/UserProfileModal"
 import { Post, usePostStore, User } from "../features/post/model/usePostStore"
-import { useShowAddDialogStore } from "../features/post/model/useShowAddDialogStore"
-import AddPostDialog from "../features/post/ui/AddPostDialog"
+import { useShowAddDialogStore } from "../features/dialog/model/useShowAddDialogStore"
+import AddPostDialog from "../features/dialog/ui/AddPostDialog"
+import { useShowEditDialogStore } from "../features/dialog/model/useShowEditPostDialogStore"
+import EditPostDialog from "../features/dialog/ui/EditPostDialog"
+import { useSelectedPostStore } from "../features/post/model/useSelectedPostStore"
 
 interface PostResponse {
   posts?: Post[]
@@ -86,12 +89,12 @@ const PostsManager = () => {
   const [total, setTotal] = useState<number>(0)
 
   // 선택한 게시물, 댓글, 사용자 상태 관리
-  const [selectedPost, setSelectedPost] = useState<Post | null>(null)
+  const selectedPost = useSelectedPostStore((state) => state.selectedPost)
+  const setSelectedPost = useSelectedPostStore((state) => state.setSelectedPost)
   const [selectedComment, setSelectedComment] = useState<Comment | null>(null)
   const [selectedUser, setSelectedUser] = useState<SelectedUser | null>(null)
 
   // URL 쿼리 파라미터 상태 관리
-  // const { selectedTag, setSelectedTag } = useSelectedTag()
   const [selectedTag, setSelectedTag] = useState<string>(queryParams.get("tag") || "")
   const [skip, setSkip] = useState<number>(parseInt(queryParams.get("skip") || "0"))
   const [limit, setLimit] = useState<number>(parseInt(queryParams.get("limit") || "10"))
@@ -106,7 +109,8 @@ const PostsManager = () => {
 
   // 대화상자 및 모달 상태 관리
   const setShowAddDialog = useShowAddDialogStore((state) => state.setShowAddDialog)
-  const [showEditDialog, setShowEditDialog] = useState<boolean>(false)
+  const setShowEditDialog = useShowEditDialogStore((state) => state.setShowEditDialog)
+
   const [showAddCommentDialog, setShowAddCommentDialog] = useState<boolean>(false)
   const [showEditCommentDialog, setShowEditCommentDialog] = useState<boolean>(false)
   const [showPostDetailDialog, setShowPostDetailDialog] = useState<boolean>(false)
@@ -209,22 +213,6 @@ const PostsManager = () => {
       console.error("태그별 게시물 가져오기 오류:", error)
     }
     setLoading(false)
-  }
-
-  // 게시물 업데이트
-  const updatePost = async () => {
-    try {
-      const response = await fetch(`/api/posts/${selectedPost?.id}`, {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(selectedPost),
-      })
-      const data = await response.json()
-      setPosts(posts.map((post) => (post.id === data.id ? data : post)))
-      setShowEditDialog(false)
-    } catch (error) {
-      console.error("게시물 업데이트 오류:", error)
-    }
   }
 
   // 게시물 삭제
@@ -624,29 +612,7 @@ const PostsManager = () => {
       </CardContent>
 
       <AddPostDialog />
-
-      {/* 게시물 수정 대화상자 */}
-      <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>게시물 수정</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <Input
-              placeholder="제목"
-              value={selectedPost?.title || ""}
-              onChange={(e) => selectedPost && setSelectedPost({ ...selectedPost, title: e.target.value })}
-            />
-            <Textarea
-              rows={15}
-              placeholder="내용"
-              value={selectedPost?.body || ""}
-              onChange={(e) => selectedPost && setSelectedPost({ ...selectedPost, body: e.target.value })}
-            />
-            <Button onClick={updatePost}>게시물 업데이트</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <EditPostDialog />
 
       {/* 댓글 추가 대화상자 */}
       <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
